+++ libraries/session/src/main/java/androidx/media3/session/MediaSession.java
@@ -58,6 +58,7 @@ import androidx.media3.common.VideoSize;
 import androidx.media3.common.util.BitmapLoader;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
+import androidx.media3.datasource.DataSourceBitmapLoader;
 import androidx.media3.session.MediaLibraryService.LibraryParams;
 import androidx.media3.session.MediaLibraryService.MediaLibrarySession;
 import com.google.common.base.Objects;
@@ -330,15 +331,14 @@ public class MediaSession {
 
     /**
      * Sets a {@link BitmapLoader} for the {@link MediaSession} to decode bitmaps from compressed
-     * binary data or load bitmaps from {@link Uri}. If not set, a {@link CacheBitmapLoader} with a
-     * {@link SimpleBitmapLoader} inside will be used.
+     * binary data or load bitmaps from {@link Uri}.
      *
      * <p>The provided instance will likely be called repeatedly with the same request, so it would
      * be best if any provided instance does some caching. Simple caching can be added to any {@link
      * BitmapLoader} implementation by wrapping it in {@link CacheBitmapLoader} before passing it to
      * this method.
      *
-     * <p>If no instance is set, a {@link CacheBitmapLoader} with a {@link SimpleBitmapLoader}
+     * <p>If no instance is set, a {@link CacheBitmapLoader} with a {@link DataSourceBitmapLoader}
      * inside will be used.
      *
      * @param bitmapLoader The bitmap loader {@link BitmapLoader}.
@@ -416,7 +416,7 @@ public class MediaSession {
     @Override
     public MediaSession build() {
       if (bitmapLoader == null) {
-        bitmapLoader = new CacheBitmapLoader(new SimpleBitmapLoader());
+        bitmapLoader = new CacheBitmapLoader(new DataSourceBitmapLoader(context));
       }
       return new MediaSession(
           context,