+++ libraries/test_session_current/src/androidTest/java/androidx/media3/session/MediaUtilsTest.java
@@ -60,6 +60,7 @@ import androidx.media3.common.Timeline;
 import androidx.media3.common.TrackGroup;
 import androidx.media3.common.Tracks;
 import androidx.media3.common.util.BitmapLoader;
+import androidx.media3.datasource.DataSourceBitmapLoader;
 import androidx.media3.session.PlayerInfo.BundlingExclusions;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
@@ -85,7 +86,7 @@ public final class MediaUtilsTest {
   @Before
   public void setUp() {
     context = ApplicationProvider.getApplicationContext();
-    bitmapLoader = new CacheBitmapLoader(new SimpleBitmapLoader());
+    bitmapLoader = new CacheBitmapLoader(new DataSourceBitmapLoader(context));
   }
 
   @Test
@@ -99,7 +100,7 @@ public final class MediaUtilsTest {
 
     MediaItem mediaItem = MediaUtils.convertToMediaItem(browserItem);
     assertThat(mediaItem.mediaId).isEqualTo(mediaId);
-    assertThat(mediaItem.mediaMetadata.title).isEqualTo(title);
+    assertThat(mediaItem.mediaMetadata.title.toString()).isEqualTo(title);
   }
 
   @Test
@@ -161,8 +162,8 @@ public final class MediaUtilsTest {
         MediaUtils.convertToMediaDescriptionCompat(mediaItem, /* artworkBitmap= */ null);
 
     assertThat(descriptionCompat.getMediaId()).isEqualTo(mediaId);
-    assertThat(descriptionCompat.getTitle()).isEqualTo(title);
-    assertThat(descriptionCompat.getDescription()).isEqualTo(description);
+    assertThat(descriptionCompat.getTitle().toString()).isEqualTo(title);
+    assertThat(descriptionCompat.getDescription().toString()).isEqualTo(description);
     assertThat(
             descriptionCompat
                 .getExtras()
@@ -203,8 +204,8 @@ public final class MediaUtilsTest {
 
   @Test
   public void convertToMediaMetadata_withTitle() {
-    CharSequence title = "title";
-    assertThat(MediaUtils.convertToMediaMetadata(title).title).isEqualTo(title);
+    String title = "title";
+    assertThat(MediaUtils.convertToMediaMetadata(title).title.toString()).isEqualTo(title);
   }
 
   @Test
@@ -268,7 +269,7 @@ public final class MediaUtilsTest {
             mediaItem.mediaMetadata,
             "mediaId",
             Uri.parse("http://www.example.com"),
-            /* durotionsMs= */ C.TIME_UNSET,
+            /* durationMs= */ C.TIME_UNSET,
             /* artworkBitmap= */ null);
 
     assertThat(
@@ -476,7 +477,7 @@ public final class MediaUtilsTest {
   @Test
   public void convertToSessionCommands_whenSessionIsNotReadyOnSdk21_disallowsRating() {
     SessionCommands sessionCommands =
-        MediaUtils.convertToSessionCommands(/* playbackState= */ null, /* isSessionReady= */ false);
+        MediaUtils.convertToSessionCommands(/* state= */ null, /* isSessionReady= */ false);
     assertThat(sessionCommands.contains(SessionCommand.COMMAND_CODE_SESSION_SET_RATING)).isFalse();
   }
 
@@ -946,7 +947,7 @@ public final class MediaUtilsTest {
     long currentPositionMs =
         MediaUtils.convertToCurrentPositionMs(
             /* playbackStateCompat= */ null,
-            /* currentMediaMetadata= */ null,
+            /* metadataCompat= */ null,
             /* timeDiffMs= */ C.TIME_UNSET);
     assertThat(currentPositionMs).isEqualTo(0);
   }
@@ -1113,7 +1114,8 @@ public final class MediaUtilsTest {
             oldPlayerInfo,
             BundlingExclusions.NONE,
             newPlayerInfo,
-            new BundlingExclusions(/* isTimelineExcluded= */ true, /* areTracksExcluded= */ true),
+            new BundlingExclusions(
+                /* isTimelineExcluded= */ true, /* areCurrentTracksExcluded= */ true),
             availableCommands);
 
     assertThat(mergeResult.first.timeline).isSameInstanceAs(oldPlayerInfo.timeline);
@@ -1155,7 +1157,8 @@ public final class MediaUtilsTest {
             oldPlayerInfo,
             BundlingExclusions.NONE,
             newPlayerInfo,
-            new BundlingExclusions(/* isTimelineExcluded= */ true, /* areTracksExcluded= */ true),
+            new BundlingExclusions(
+                /* isTimelineExcluded= */ true, /* areCurrentTracksExcluded= */ true),
             availableCommands);
 
     assertThat(mergeResult.first.timeline).isSameInstanceAs(Timeline.EMPTY);
@@ -1197,7 +1200,8 @@ public final class MediaUtilsTest {
             oldPlayerInfo,
             BundlingExclusions.NONE,
             newPlayerInfo,
-            new BundlingExclusions(/* isTimelineExcluded= */ true, /* areTracksExcluded= */ true),
+            new BundlingExclusions(
+                /* isTimelineExcluded= */ true, /* areCurrentTracksExcluded= */ true),
             availableCommands);
 
     assertThat(mergeResult.first.timeline).isSameInstanceAs(oldPlayerInfo.timeline);